change user service, add auth service and intersepter
