+++ src/api/api.ts
@@ -1,0 +1,21 @@
+import { getAccessToken } from '@/utils/utils';
+import axios, { InternalAxiosRequestConfig } from 'axios';
+
+const apiClient = axios.create({
+    baseURL: process.env.IS_DEV === 'development' ? '/api' : process.env.API_URL,
+    headers: {
+        'Content-Type': 'application/json',
+    },
+});
+
+apiClient.interceptors.request.use((config: InternalAxiosRequestConfig) => {
+    const token = getAccessToken();
+    if (token) {
+        config.headers.Authorization = `Bearer ${token}`
+    }
+    return config;
+}, (error) => {
+    return Promise.reject(error);
+});
+
+export default apiClient;
