+++ src/components/blocks/footer/Footer.tsx
@@ -5,7 +5,7 @@ import CustomButton from "@/components/ui/custom-button/CustomButton";
 const Footer: React.FC = () => {
   return (
     <Container>
-      <div className="flex justify-between items-center py-4 pt-6 pb-6">
+      <div className="flex justify-between items-center py-4 pt-8 pb-6">
         <img src="" alt="" className="w-[100px] h-[30px] mr-10" />
 
         <HeaderMenu />
